moved validatetokenresponse to domain
