+++ lib/features/authentication/domain/validate_token_response.dart
